Moved set up @zotmeal/ui to work with imports
