+++ apps/expo/tsconfig.json
@@ -17,7 +17,7 @@
     "*.js",
     ".expo/types/**/*.ts",
     "expo-env.d.ts"
-  ],
+, "../../packages/ui/src/home_view/Home.tsx"  ],
   "exclude": [
     "node_modules"
   ]
