+++ packages/ui/src/home_view/example_data.ts
@@ -1,0 +1,347 @@
+// TODO: Replace with real data
+export const brandywineData = {
+    date: "1/22/2022",
+    period: "breakfast",
+    restaurant: "brandywine",
+    stations: [
+      {
+        id: "111",
+        name: "Honeycakes/Bakery",
+        categories: [
+          {
+            name: "Desserts",
+            dishes: [
+              {
+                name: "Blueberry Scone",
+                id: "123",
+                description: "Item Description",
+                nutritionInfo: {
+                  calories: "300",
+                  dietaryFiber: "2g",
+                  protein: "10g",
+                  sugars: "5g",
+                  totalCarbohydrates: "30g",
+                  totalFat: "15g",
+                  transFat: "0g",
+                  saturatedFat: "5g",
+                  sodium: "300mg",
+                  cholesterol: "50mg",
+                  vitaminA: "10%",
+                  vitaminC: "10%",
+                  calcium: "10%",
+                  iron: "10%",
+                },
+              },
+              {
+                name: "Breakfast Muffin",
+                id: "312",
+                description: "Item Description",
+                nutritionInfo: {
+                  calories: "300",
+                  dietaryFiber: "2g",
+                  protein: "10g",
+                  sugars: "5g",
+                  totalCarbohydrates: "30g",
+                  totalFat: "15g",
+                  transFat: "0g",
+                  saturatedFat: "5g",
+                  sodium: "300mg",
+                  cholesterol: "50mg",
+                  vitaminA: "10%",
+                  vitaminC: "10%",
+                  calcium: "10%",
+                  iron: "10%",
+                },
+              },
+            ],
+          },
+          {
+            name: "Sides",
+            dishes: [
+              {
+                name: "Side 1",
+                id: "444",
+                description: "Item Description",
+                nutritionInfo: {
+                  calories: "300",
+                  dietaryFiber: "2g",
+                  protein: "10g",
+                  sugars: "5g",
+                  totalCarbohydrates: "30g",
+                  totalFat: "15g",
+                  transFat: "0g",
+                  saturatedFat: "5g",
+                  sodium: "300mg",
+                  cholesterol: "50mg",
+                  vitaminA: "10%",
+                  vitaminC: "10%",
+                  calcium: "10%",
+                  iron: "10%",
+                },
+              },
+              {
+                name: "Side 2",
+                id: "555",
+                description: "Item Description",
+                nutritionInfo: {
+                  calories: "300",
+                  dietaryFiber: "2g",
+                  protein: "10g",
+                  sugars: "5g",
+                  totalCarbohydrates: "30g",
+                  totalFat: "15g",
+                  transFat: "0g",
+                  saturatedFat: "5g",
+                  sodium: "300mg",
+                  cholesterol: "50mg",
+                  vitaminA: "10%",
+                  vitaminC: "10%",
+                  calcium: "10%",
+                  iron: "10%",
+                },
+              },
+            ],
+          },
+        ],
+      },
+      {
+        id: "222",
+        name: "Hearth/Pizza",
+        categories: [
+          {
+            name: "Entrees",
+            dishes: [
+              {
+                name: "Blueberry Scone",
+                id: "123",
+                description: "Item Description",
+                nutritionInfo: {
+                  calories: "300",
+                  dietaryFiber: "2g",
+                  protein: "10g",
+                  sugars: "5g",
+                  totalCarbohydrates: "30g",
+                  totalFat: "15g",
+                  transFat: "0g",
+                  saturatedFat: "5g",
+                  sodium: "300mg",
+                  cholesterol: "50mg",
+                  vitaminA: "10%",
+                  vitaminC: "10%",
+                  calcium: "10%",
+                  iron: "10%",
+                },
+                dietRestriction: {
+                  isVegan: true,
+                  isVegetarian: true,
+                  isGlutenFree: true,
+                  isHalal: true,
+                },
+              },
+              {
+                name: "Breakfast Muffin",
+                id: "312",
+                description: "Item Description",
+                nutritionInfo: {
+                  calories: "300",
+                  dietaryFiber: "2g",
+                  protein: "10g",
+                  sugars: "5g",
+                  totalCarbohydrates: "30g",
+                  totalFat: "15g",
+                  transFat: "0g",
+                  saturatedFat: "5g",
+                  sodium: "300mg",
+                  cholesterol: "50mg",
+                  vitaminA: "10%",
+                  vitaminC: "10%",
+                  calcium: "10%",
+                  iron: "10%",
+                },
+                dietRestriction: {
+                  isVegan: true,
+                  isVegetarian: true,
+                  isGlutenFree: true,
+                  isHalal: true,
+                },
+              },
+            ],
+          },
+        ],
+      },
+    ]
+  };
+  
+  // TODO: Replace with real data
+  export const anteateryData = {
+    date: "1/22/2022",
+    period: "breakfast",
+    restaurant: "anteatery",
+    stations: [
+      {
+        id: "111",
+        name: "Oven",
+        categories: [
+          {
+            name: "Entrees",
+            dishes: [
+              {
+                name: "Blueberry Scone",
+                id: "123",
+                description: "Item Description",
+                nutritionInfo: {
+                  calories: "300",
+                  dietaryFiber: "2g",
+                  protein: "10g",
+                  sugars: "5g",
+                  totalCarbohydrates: "30g",
+                  totalFat: "15g",
+                  transFat: "0g",
+                  saturatedFat: "5g",
+                  sodium: "300mg",
+                  cholesterol: "50mg",
+                  vitaminA: "10%",
+                  vitaminC: "10%",
+                  calcium: "10%",
+                  iron: "10%",
+                },
+              },
+              {
+                name: "Breakfast Muffin",
+                id: "312",
+                description: "Item Description",
+                nutritionInfo: {
+                  calories: "300",
+                  dietaryFiber: "2g",
+                  protein: "10g",
+                  sugars: "5g",
+                  totalCarbohydrates: "30g",
+                  totalFat: "15g",
+                  transFat: "0g",
+                  saturatedFat: "5g",
+                  sodium: "300mg",
+                  cholesterol: "50mg",
+                  vitaminA: "10%",
+                  vitaminC: "10%",
+                  calcium: "10%",
+                  iron: "10%",
+                },
+              },
+            ],
+          },
+          {
+            name: "Sides",
+            dishes: [
+              {
+                name: "Side 1",
+                id: "444",
+                description: "Item Description",
+                nutritionInfo: {
+                  calories: "300",
+                  dietaryFiber: "2g",
+                  protein: "10g",
+                  sugars: "5g",
+                  totalCarbohydrates: "30g",
+                  totalFat: "15g",
+                  transFat: "0g",
+                  saturatedFat: "5g",
+                  sodium: "300mg",
+                  cholesterol: "50mg",
+                  vitaminA: "10%",
+                  vitaminC: "10%",
+                  calcium: "10%",
+                  iron: "10%",
+                },
+              },
+              {
+                name: "Side 2",
+                id: "555",
+                description: "Item Description",
+                nutritionInfo: {
+                  calories: "300",
+                  dietaryFiber: "2g",
+                  protein: "10g",
+                  sugars: "5g",
+                  totalCarbohydrates: "30g",
+                  totalFat: "15g",
+                  transFat: "0g",
+                  saturatedFat: "5g",
+                  sodium: "300mg",
+                  cholesterol: "50mg",
+                  vitaminA: "10%",
+                  vitaminC: "10%",
+                  calcium: "10%",
+                  iron: "10%",
+                },
+              },
+            ],
+          },
+        ],
+      },
+      {
+        id: "222",
+        name: "Hearth/Pizza",
+        categories: [
+          {
+            name: "Bakery",
+            dishes: [
+              {
+                name: "Blueberry Scone",
+                id: "123",
+                description: "Item Description",
+                nutritionInfo: {
+                  calories: "300",
+                  dietaryFiber: "2g",
+                  protein: "10g",
+                  sugars: "5g",
+                  totalCarbohydrates: "30g",
+                  totalFat: "15g",
+                  transFat: "0g",
+                  saturatedFat: "5g",
+                  sodium: "300mg",
+                  cholesterol: "50mg",
+                  vitaminA: "10%",
+                  vitaminC: "10%",
+                  calcium: "10%",
+                  iron: "10%",
+                },
+                dietRestriction: {
+                  isVegan: true,
+                  isVegetarian: true,
+                  isGlutenFree: true,
+                  isHalal: true,
+                },
+              },
+              {
+                name: "Breakfast Muffin",
+                id: "312",
+                description: "Item Description",
+                nutritionInfo: {
+                  calories: "300",
+                  dietaryFiber: "2g",
+                  protein: "10g",
+                  sugars: "5g",
+                  totalCarbohydrates: "30g",
+                  totalFat: "15g",
+                  transFat: "0g",
+                  saturatedFat: "5g",
+                  sodium: "300mg",
+                  cholesterol: "50mg",
+                  vitaminA: "10%",
+                  vitaminC: "10%",
+                  calcium: "10%",
+                  iron: "10%",
+                },
+                dietRestriction: {
+                  isVegan: true,
+                  isVegetarian: true,
+                  isGlutenFree: true,
+                  isHalal: true,
+                },
+              },
+            ],
+          },
+        ],
+      },
+    ]
+  };
