+++ packages/ui/src/home_view/index.ts
@@ -1,0 +1,1 @@
+export { Home } from "./Home"
